+++ src/delta.rs
@@ -4,7 +4,7 @@ use super::polynomial;
 use super::term_builder;
 
 /// (1-q)^24 * (1-q^2)^24 * .. * (1-q^order)^24
-pub fn delta1(order: i64) -> polynomial::Polynomial {
+pub fn delta1(order: i32) -> polynomial::Polynomial {
     let one = polynomial::Polynomial::one();
     let mut pol = polynomial::Polynomial::one();
     for n in num_iter::range(1, order + 1) {
@@ -17,7 +17,7 @@ pub fn delta1(order: i64) -> polynomial::Polynomial {
 }
 
 // 1/d = 1 + (1-d) + (1-d)^2 + ...
-pub fn delta1_inverse(order: i64) -> polynomial::Polynomial {
+pub fn delta1_inverse(order: i32) -> polynomial::Polynomial {
     let a = polynomial::Polynomial::one() - delta1(order);
     let mut pol = polynomial::Polynomial::one();
     for n in num_iter::range(1, order + 1) {
